Update css_main
Positions are not added, this will be done once all html pages are done. 

But this can just be the main css which is just all of the font and colours provided. 

Sizing of elements will be added once all html pages are done so I can test how big... (in px)